% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approved_model.R
\name{approved_model}
\alias{approved_model}
\title{copy the previous approved model as the new "base" model}
\usage{
approved_model(base_year, base_model, year)
}
\arguments{
\item{base_year}{year of the base model}

\item{base_model}{name of the base model (folder name)}

\item{year}{current year}
}
\value{

}
\description{
copy of all base model files, also add a "base_model.txt" file that identifies the model used
}
\examples{
approved_model(base_year = 2018, base_model = "m2", year = 2020)
}
