% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_plots.R
\name{base_plots}
\alias{base_plots}
\title{all basic plost forsafe doc}
\usage{
base_plots(year, model, model_name, rec_age)
}
\arguments{
\item{year}{is the model year}

\item{model}{is the folder the model is in}

\item{model_name}{the name of the tpl file (no extension)}

\item{rec_age}{is the recruitment age}
}
\value{

}
\description{
all basic plost forsafe doc
}
