% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_catch.R
\name{clean_catch}
\alias{clean_catch}
\title{clean up catch data}
\usage{
clean_catch(year = 2020, TAC = c(3333, 2222, 1111), old_catch = NULL)
}
\arguments{
\item{year}{- year of assessment}

\item{TAC}{- last three TAC in form: c(year-3, year-2, year-1)}

\item{old_catch}{- if catch is frozen place the file in user_input folder}
}
\value{

}
\description{
clean up catch data
}
\examples{
clean_catch(year, TAC, old_catch = )
}
