% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohn.R
\name{mohn}
\alias{mohn}
\title{Calculate Mohn's rho "mean version"}
\usage{
mohn(year, model, n_retro = 10)
}
\arguments{
\item{year}{assessment year}

\item{model}{folder model is in}

\item{n_retro}{number of retrospective analyses to do (default = 10)}
}
\value{

}
\description{
Calculate Mohn's rho "mean version"
}
\examples{
mohn(year, model)
}
