% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_re.R
\name{plot_re}
\alias{plot_re}
\title{plot random effects model by GOA area}
\usage{
plot_re(location)
}
\arguments{
\item{location}{where the files live e.g., "Y:/ABL_MESA/SAFES2020/Apportionment/Plot data - NR"}
}
\value{

}
\description{
plot random effects model by GOA area
}
\examples{
\dontrun{
plot_re(location)
}

}
