% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_retro.R
\name{plot_retro}
\alias{plot_retro}
\title{retrospective plot}
\usage{
plot_retro(year, model, n_retro = 10)
}
\arguments{
\item{year}{assessment year}

\item{model}{folder the model is in}

\item{n_retro}{default is 10}
}
\value{

}
\description{
retrospective plot
}
\examples{
plot_retro(year, model)
}
