% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_retro_survey.R
\name{plot_retro_survey}
\alias{plot_retro_survey}
\title{retrospective survey plot}
\usage{
plot_retro_survey(year, model, n_retro = 10)
}
\arguments{
\item{year}{assessment year}

\item{model}{folder the model is in}

\item{n_retro}{default is 10}
}
\value{

}
\description{
retrospective survey plot
}
\examples{
plot_retro_survey(year, model)
}
