% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_retro.R
\name{run_retro}
\alias{run_retro}
\title{Run retrospectives}
\usage{
run_retro(
  year,
  model,
  tpl_name,
  n_retro = 10,
  admb_home = NULL,
  mcon = NULL,
  mcmc = 5e+05,
  mcsave = 100
)
}
\arguments{
\item{year}{assessment year}

\item{model}{folder model is in}

\item{tpl_name}{name of tpl file}

\item{n_retro}{how many retros? default is 10}

\item{admb_home}{where is admb on your computer}

\item{mcon}{currently nulll}

\item{mcmc}{how many mcmcs to run}

\item{mcsave}{how many mcmcs to save}
}
\value{

}
\description{
Run retrospectives
}
